step2 : 피드백 반영 - Lotto.kt 안의 클래스 파일 분리, 로또 생성시 IntRange 적용
